requirements.txt
absl-py==2.0.0
astunparse==1.6.3
autopep8==2.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.1.0
chex==0.1.85
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
etils==1.6.0
Flask==2.3.2
flatbuffers==23.5.26
flax==0.7.5
fonttools==4.38.0
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.2
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.3
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.4
packaging==23.2
pandas==2.1.4
Pillow==9.4.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
Pygments==2.17.2
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scipy==1.11.4
six==1.16.0
soupsieve==2.4.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.45
termcolor==2.4.0
toolz==0.12.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.17.0
